Rename C# project to GuildedRose
