+++ csharpcore/GuildedRose/GildedRose.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace csharpcore
+namespace GuildedRose
 {
     public class GildedRose
     {
