+++ csharpcore/GuildedRose/GuildedRose.csproj
@@ -3,7 +3,7 @@
   <PropertyGroup>
     <OutputType>Exe</OutputType>
     <TargetFramework>netcoreapp3.1</TargetFramework>
-    <StartupObject>csharpcore.Program</StartupObject>
+    <StartupObject>GuildedRose.Program</StartupObject>
   </PropertyGroup>
 
   <ItemGroup>
