+++ csharpcore/GuildedRose/Program.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 
-namespace csharpcore
+namespace GuildedRose
 {
     public class Program
     {
